Create .clang-format file in workspace dir on bundle start
This ensures that the user has always a default formatter similar to
the CDTs default K&R format style.
