+++ bundles/org.eclipse.cdt.lsp.clangd/src/org/eclipse/cdt/lsp/clangd/utils/.clang-format-ws
@@ -1,4 +1,4 @@
-# We'll use defaults from the LLVM style, but with some modifications that it's close to the CDT K&R style.
+# We'll use defaults from the LLVM style, but with some modifications so that it's close to the CDT K&R style.
 BasedOnStyle: LLVM
 UseTab: Always
 IndentWidth: 4
